refactor: 기록 관련 도메인 배치
