+++ src/main/kotlin/com/yapp/yapp/record/domain/point/RunningPoint.kt
@@ -1,9 +1,9 @@
 package com.yapp.yapp.record.domain.point
 
 import com.yapp.yapp.common.TimeProvider
+import com.yapp.yapp.record.domain.Pace
+import com.yapp.yapp.record.domain.converter.PaceConverter
 import com.yapp.yapp.record.domain.record.RunningRecord
-import com.yapp.yapp.running.domain.Pace
-import com.yapp.yapp.running.domain.converter.PaceConverter
 import jakarta.persistence.Column
 import jakarta.persistence.Convert
 import jakarta.persistence.Entity
